Issue #28661 split data.internal bundle to prepare for dataContainer
