+++ dev/com.ibm.websphere.appserver.features/visibility/public/data-1.0/io.openliberty.data-1.0.feature
@@ -20,6 +20,7 @@ Subsystem-Name: Jakarta Data 1.0
   io.openliberty.jakarta.data-1.0
 -bundles=\
   io.openliberty.data.internal,\
+  io.openliberty.data.internal.beandef,\
   io.openliberty.data.1.0.internal
 kind=beta
 edition=base
