+++ dev/com.ibm.websphere.appserver.features/visibility/public/data-1.1/io.openliberty.data-1.1.feature
@@ -27,6 +27,7 @@ Subsystem-Name: Jakarta Data 1.1
 -bundles=\
   io.openliberty.data; location:="dev/api/ibm/,lib/",\
   io.openliberty.data.internal,\
+  io.openliberty.data.internal.beandef,\
   io.openliberty.data.1.1.internal
 -files=dev/api/ibm/javadoc/io.openliberty.data_1.1-javadoc.zip
 kind=noship
