+++ dev/io.openliberty.data.internal.beandef/.gitignore
@@ -1,0 +1,2 @@
+/bin/
+/generated/
