+++ dev/io.openliberty.data.internal.beandef/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2022,2024 IBM Corporation and others.
+# Copyright (c) 2024 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -18,19 +18,18 @@ javac.target: 17
 
 Require-Capability: osgi.ee; filter:="(&(osgi.ee=JavaSE)(version=17))"
 
-Bundle-Name: Jakarta Data Version-dependent Function
-Bundle-SymbolicName: io.openliberty.data.internal
-Bundle-Description: Coordinates differences between versions of Jakarta Data; version=${bVersion}
+Bundle-Name: Jakarta Data Bean Defining Annotations
+Bundle-SymbolicName: io.openliberty.data.internal.beandef
+Bundle-Description: Declares bean defining annotations for Jakarta Data; version=${bVersion}
 
 WS-TraceGroup: data
 
-Export-Package: \
-  io.openliberty.data.internal.version.*
-
 Private-Package: \
-  io.openliberty.data.internal.resources.*
+  io.openliberty.data.internal.beandef.*
 
-instrument.classesExcludes: io/openliberty/data/internal/resources/*.class
+-cdiannotations:
+
+-dsannotations: io.openliberty.data.internal.beandef.BeanDefiningAnnotationMetadata
 
 -buildpath: \
   com.ibm.websphere.appserver.spi.logging,\
@@ -39,4 +38,6 @@ instrument.classesExcludes: io/openliberty/data/internal/resources/*.class
   com.ibm.websphere.org.osgi.service.component,\
   com.ibm.wsspi.org.osgi.service.component.annotations,\
   com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
+  com.ibm.ws.cdi.interfaces.jakarta;version=latest, \
+  io.openliberty.jakarta.cdi.4.0,\
   io.openliberty.jakarta.data.1.0
