+++ dev/io.openliberty.data.internal.beandef/src/io/openliberty/data/internal/beandef/package-info.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2022,2023 IBM Corporation and others.
+ * Copyright (c) 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -14,7 +14,7 @@
  * @version 1.0
  */
 @org.osgi.annotation.versioning.Version("1.0")
-@TraceOptions(traceGroup = "data", messageBundle = "io.openliberty.data.internal.resources.CWWKDMessages")
-package io.openliberty.data.internal.cdi;
+@TraceOptions(traceGroup = "data")
+package io.openliberty.data.internal.beandef;
 
 import com.ibm.websphere.ras.annotation.TraceOptions;
